+++ vlo-monitor/src/main/java/eu/clarin/cmdi/vlo/monitor/service/IndexService.java
@@ -8,6 +8,6 @@ import java.util.Map;
  */
 public interface IndexService {
     
-    Map<String, Integer> getValueCounts(String facet);
+    Map<String, Long> getValueCounts(String facet);
     
 }
